UPDIS-557 rtx接口实现 需要现场测试

+++ updis/res_users.py
@@ -6,6 +6,7 @@ from openerp import SUPERUSER_ID
 
 from openerp.osv import osv, fields
 from up_tools.rtxlib import RtxClient
+from openerp.tools.translate import _
 
 
 class user_device(osv.osv):
@@ -21,6 +22,7 @@ class res_users(osv.osv):
     _inherit = "res.users"
     _bigAntClient = BigAntClient()
     _rtx_client = RtxClient()
+    _base_dep = u'规划设计一所'
     _columns = {
         'sign_image': fields.binary("Sign Image",
                                     help="This field holds the image used as siganture for this contact"),
@@ -47,36 +49,70 @@ class res_users(osv.osv):
                               'mobile_phone', 'work_location', 'interest', 'practice', 'person_resume', 'home_phone',
                               'devices', ]
 
+    def _is_need_rtx_sync(self, cr, uid, context):
+        return self.pool.get('ir.config_parameter').get_param(cr, 1, 'bigant.password_sync') == 'True'
+
+    def _is_rtx_user_exist(self, name):
+        return self._rtx_client.get_client().IsUserExist(userName=name, key=self._rtx_client.rtx_key)
+
     def write_rtx_user(self, cr, uid, ids, vals, context=None):
-        client = self._rtx_client.get_client()
-        for user in self.browse(cr, uid, ids, context):
-            params = {'userName': user.login,
+        if self._is_need_rtx_sync(cr, uid, context):
+            client = self._rtx_client.get_client()
+            for user in self.browse(cr, uid, ids, context):
+                params = {'userName': user.login,
+                          # 'userPwd':1,
+                          'DeptName': self._base_dep,
+                          'ChsName': user.name or '',
+                          'IGender': 0,
+                          'Cell': user.work_phone or '',
+                          'Email': user.work_email or '',
+                          'Phone': user.mobile_phone or '',
+                          # 'Position': '',
+                          'AuthTYpe': 0,
+                          'key': self._rtx_client.rtx_key}
+                client.EditUser(**params)
+                if user.active is False:
+                    self.unlink_rtx_user(cr, uid, ids, context)
+                elif not self._is_rtx_user_exist(user.login):
+                    values = {
+                        'login': user.login or '',
+                        'name': user.name or '',
+                        'work_phone': user.work_phone or '',
+                        'work_email': user.work_email or '',
+                        'mobile_phone': user.mobile_phone or '',
+                    }
+                    self.add_rtx_user(cr, uid, values, context)
+
+    def add_rtx_user(self, cr, uid, vals, context=None):
+        if self._is_need_rtx_sync(cr, uid, context):
+            params = {'userName': vals.get('login', ''),
                       # 'userPwd':1,
-                      # 'DeptName': u'临时',
-                      'ChsName': user.name,
-                      'IGender': 0,
-                      'Cell': user.work_phone,
-                      'Email': user.work_email,
-                      'Phone': user.mobile_phone,
+                      'DeptName': self._base_dep,
+                      'ChsName': vals.get('name', ''),
+                      'IGender': 1,
+                      'Cell': vals.get('work_phone', ''),
+                      'Email': vals.get('work_email', ''),
+                      'Phone': vals.get('mobile_phone', ''),
                       # 'Position':'',
                       'AuthTYpe': 0,
                       'key': self._rtx_client.rtx_key}
-            client.EditUser(**params)
-            if user.is_active is False:
+            if self._is_rtx_user_exist(vals.get('login')):
+                raise osv.except_osv(_('Warning!'), _('当前登录名已经在RTX中使用.'))
+            else:
+                self._rtx_client.get_client().AddUser(**params)
+
+    def unlink_rtx_user(self, cr, uid, ids, context):
+        if self._is_need_rtx_sync(cr, uid, context):
+            client = self._rtx_client.get_client()
+            for user in self.browse(cr, uid, ids, context):
                 try:
-                    self._rtx_client.get_client().DeleteUser(userName=user.login, key=self._rtx_client.rtx_key)
+                    client.DeleteUser(userName=user.login, key=self._rtx_client.rtx_key)
                 except Exception:
                     pass
 
-
     def unlink(self, cr, uid, ids, context=None):
-        client = self._rtx_client.get_client()
-        for user in self.browse(cr, uid, ids, context):
-            try:
-                client.DeleteUser(userName=user.login, key=self._rtx_client.rtx_key)
-            except Exception:
-                pass
-        super(res_users, self).unlink(cr, uid, ids, context)
+        self.unlink_rtx_user(cr, uid, ids, context)
+        return super(res_users, self).unlink(cr, uid, ids, context)
 
     def write(self, cr, uid, ids, values, context=None):
         if not hasattr(ids, '__iter__'):
@@ -87,8 +123,8 @@ class res_users(osv.osv):
                     break
             else:
                 if 'company_id' in values:
-                    if not (values['company_id'] in self.read(cr, SUPERUSER_ID, uid, ['company_ids'], context=context)[
-                        'company_ids']):
+                    if not (values['company_id'] in self.read(cr, SUPERUSER_ID, uid, ['company_ids'],
+                                                              context=context)['company_ids']):
                         del values['company_id']
                 uid = 1  # safe fields only, so we write as super-user to bypass access rights
         else:
@@ -112,24 +148,9 @@ class res_users(osv.osv):
                 if id in self._uid_cache[db]:
                     del self._uid_cache[db][id]
         self.context_get.clear_cache(self)
+
         self.write_rtx_user(cr, 1, ids, values, context)
         return res
-
-    def add_rtx_user(self, cr, uid, vals, context=None):
-        params = {'userName': vals.get('login', ''),
-                  # 'userPwd':1,
-                  'DeptName': u'临时',
-                  'ChsName': vals.get('name', ''),
-                  'IGender': 1,
-                  'Cell': vals.get('work_phone', ''),
-                  'Email': vals.get('work_email', ''),
-                  'Phone': vals.get('mobile_phone', ''),
-                  # 'Position':'',
-                  'AuthTYpe': 0,
-                  'key': self._rtx_client.rtx_key}
-        self._rtx_client.get_client().AddUser(**params)
-        # self._rtx_client.client.AddUser(userName='1111111', userPwd='111', DeptName='', ChsName='', IGender=1, Cell='', Email='', Phone='',
-        # Position='', AuthTYpe=0, key='tianvService2014')
 
     def create(self, cr, uid, vals, context=None):
         self.add_rtx_user(cr, uid, vals, context)
@@ -137,7 +158,6 @@ class res_users(osv.osv):
             return super(res_users, self).create(cr, 1, vals, context)
         else:
             return super(res_users, self).create(cr, uid, vals, context)
-
 
     def on_change_login(self, cr, uid, ids, login, big_ant_name, context=None):
         ret = {'value': {}}
@@ -173,7 +193,7 @@ class res_users(osv.osv):
     def change_password(self, cr, uid, old_passwd, new_passwd, context=None):
         result = super(res_users, self).change_password(cr, uid, old_passwd, new_passwd, context=context)
         user = self.browse(cr, 1, uid, context)
-        if self.pool.get('ir.config_parameter').get_param(cr, 1, 'bigant.password_sync') == 'True':
+        if self.pool.get('ir.config_parameter').get_param(cr, 1, 'bigant.password_sync') == 'True' and self._is_rtx_user_exist(user.login):
             params = {
                 'userName': user.login,
                 'password': new_passwd,
@@ -186,13 +206,15 @@ class res_users(osv.osv):
 class ChangePasswordUser(osv.TransientModel):
     _inherit = 'change.password.user'
     _bigAntClient = BigAntClient()
+    _rtx_client = RtxClient()
 
     def change_password_button(self, cr, uid, ids, context=None):
         for user in self.browse(cr, uid, ids, context=context):
             self.pool.get('res.users').write(cr, uid, user.user_id.id, {'password': user.new_passwd})
-            if self.pool.get('ir.config_parameter').get_param(cr, 1, 'bigant.password_sync') == 'True':
+            if self.pool.get('ir.config_parameter').get_param(cr, 1, 'bigant.password_sync') == 'True' and \
+                    self._rtx_client.get_client().IsUserExist(userName=user.user_id.login, key=self._rtx_client.rtx_key):
                 params = {
-                    'userName': user.login,
+                    'userName': user.user_id.login,
                     'password': user.new_passwd,
                     'key': self._rtx_client.rtx_key,
                 }
